refactor: deprecate NestedBaseTrajectory and NestedRosStampedFeature
- Deleted legacy nested trajectory dataclasses (`NestedBaseTrajectory`,
  `NestedRosStampedFeature`) since nested logic is now supported by `AbstractTrajectoryCommon`.
- Updated ROS namespace imports and related references to reflect
  the removal of deprecated dataclasses.
- Adjusted tests, documentation, and examples to align with the reduced
  class structure.

Issue TCT-87
