+++ documentation/timestamp_utilities.md
@@ -155,28 +155,30 @@ For `AbstractTrajectoryStampedFeaturesBag` containers (e.g., extracted from ROS 
 Retrieve trajectory data within a specified timestamp range:
 
 ```python
+import trajectory_container_tools.dataclasses.ros_msgs.ackermann_msgs_dataclass
+import trajectory_container_tools.dataclasses.ros_msgs.nav_msgs_dataclass
 import trajectory_container_tools as tct
 
 # Extract trajectory features data from ROS bag
 trajectory_features_bag = tct.extractor.from_rosbag(
-    rosbag_path=rosbag_path,
-    features_config={
-        '/odom': tct.dataclasses.NavMsgsOdometry,
-        '/cmd': tct.dataclasses.AckermannMsgsAckermannDriveStamped,
-    },
-    chunk_on="/cmd",
-)
+        rosbag_path=rosbag_path,
+        features_config={
+                '/odom': trajectory_container_tools.dataclasses.ros_msgs.nav_msgs_dataclass.NavMsgsOdometry,
+                '/cmd':  trajectory_container_tools.dataclasses.ros_msgs.ackermann_msgs_dataclass.AckermannMsgsAckermannDriveStamped,
+                },
+        chunk_on="/cmd",
+        )
 
 # Extract data within specific timestamp window
 start_time = 1695601812731601521
 stop_time = 1695601815000000000
 
 windowed_data = trajectory_features_bag.get_timestamps(
-    start=start_time,
-    stop=stop_time,
-    startpoint=True,   # Include start timestamp
-    endpoint=False     # Exclude end timestamp
-)
+        start=start_time,
+        stop=stop_time,
+        startpoint=True,  # Include start timestamp
+        endpoint=False  # Exclude end timestamp
+        )
 
 print(f"Original chunks: {trajectory_features_bag.chunks_total}")
 print(f"Windowed chunks: {windowed_data.chunks_total}")
@@ -242,18 +244,20 @@ if limits.first <= my_timestamp and my_timestamp <= limits.last:
 ### Example 1: Synchronize Multi-Sensor Data
 
 ```python
+import trajectory_container_tools.dataclasses.ros_msgs.ackermann_msgs_dataclass
+import trajectory_container_tools.dataclasses.ros_msgs.nav_msgs_dataclass
 import trajectory_container_tools as tct
 
 # Extract multi-sensor data
 robot_data = tct.extractor.from_rosbag(
-    rosbag_path=rosbag_path,
-    features_config={
-        "/odom": tct.dataclasses.NavMsgsOdometry,
-        "/imu": tct.dataclasses.SensorMsgsImu,
-        "/cmd": tct.dataclasses.AckermannMsgsAckermannDriveStamped,
-    },
-    chunk_on="/cmd",
-)
+        rosbag_path=rosbag_path,
+        features_config={
+                "/odom": trajectory_container_tools.dataclasses.ros_msgs.nav_msgs_dataclass.NavMsgsOdometry,
+                "/imu":  tct.dataclasses.SensorMsgsImu,
+                "/cmd":  trajectory_container_tools.dataclasses.ros_msgs.ackermann_msgs_dataclass.AckermannMsgsAckermannDriveStamped,
+                },
+        chunk_on="/cmd",
+        )
 
 # Get timestamps for each sensor
 odom_stamps = robot_data.topic_odom.header.timestamps
@@ -261,12 +265,12 @@ cmd_stamps = robot_data.topic_cmd.header.timestamps
 
 # Find synchronized timestamps
 for cmd_stamp in cmd_stamps.stamps[:10]:  # First 10 commands
-    # Find nearest future odometry reading
-    if odom_stamps.is_timestamps_in_bounds(cmd_stamp):
-        nearest_odom_stamp = odom_stamps.get_nearest_stamp(
+  # Find nearest future odometry reading
+  if odom_stamps.is_timestamps_in_bounds(cmd_stamp):
+    nearest_odom_stamp = odom_stamps.get_nearest_stamp(
             cmd_stamp, future=True, include=False
-        )
-        print(f"Command at {cmd_stamp} -> Odom at {nearest_odom_stamp}")
+            )
+    print(f"Command at {cmd_stamp} -> Odom at {nearest_odom_stamp}")
 
 ```
