+++ src/trajectory_container_tools/__init__.py
@@ -11,10 +11,10 @@ A library for managing trajectory-related data with support for:
 
 Quick Start:
 
->>> import trajectory_container_tools as tct
+>>>import trajectory_container_tools.dataclasses.ros_msgs.nav_msgs_dataclass import trajectory_container_tools as tct
 >>>
 >>> features_config = {
->>>     '/odom': tct.dataclasses.NavMsgsOdometry,
+>>>     '/odom': trajectory_container_tools.dataclasses.ros_msgs.nav_msgs_dataclass.NavMsgsOdometry,
 >>>     '/tf': tct.dataclasses.Tf2MsgsTFMessage,
 >>> }
 >>>
@@ -25,7 +25,7 @@ Quick Start:
 >>> data = tct.extractor.from_dataframe(df, features_config)
 >>>
 >>> # Access dataclasses
->>> odom_class = tct.dataclasses.ros_msgs.stamped_dataclass.NavMsgsOdometry
+>>> odom_class = trajectory_container_tools.dataclasses.ros_msgs.nav_msgs_dataclass.NavMsgsOdometry
 >>>
 >>> # Utilities
 >>> tct.extractor.check_bag_topics(rosbag_path)
@@ -40,7 +40,7 @@ from .version import __version__
 from .dataclasses.core.abstract_trajectory_feature_dataclass import (
     AbstractTrajectoryFeature,
 )
-from .dataclasses.core.abstract_array_trajectory_dataclass import (
+from .dataclasses.core.abstract_trajectory_array_dataclass import (
     AbstractTrajectoryArray,
 )
 from .dataclasses.core.abstract_trajectory_features_bag_dataclass import (
@@ -51,8 +51,7 @@ from .dataclasses.core.abstract_trajectory_stamped_features_bag_dataclass import
 )
 from .dataclasses.core.base_trajectory_dataclass import (
     BaseTrajectoryFeature,
-    NestedBaseTrajectory,
-    BaseTrajectoryArray,
+    BaseTrajectoryFeatureArray,
 )
 
 from .utils.containers_sanity_checks import containers_timestep_alignment_sanity_check
@@ -78,8 +77,7 @@ __all__ = [
     "AbstractTrajectoryFeaturesBag",
     "AbstractTrajectoryStampedFeaturesBag",
     "BaseTrajectoryFeature",
-    "NestedBaseTrajectory",
-    "BaseTrajectoryArray",
+    "BaseTrajectoryFeatureArray",
     # Namespaces
     "dataclasses",
     "factory",
