+++ src/trajectory_container_tools/dataclasses/core/__init__.py
@@ -1,12 +1,11 @@
 # coding=utf-8
 from .abstract_trajectory_feature_dataclass import AbstractTrajectoryFeature
-from .abstract_array_trajectory_dataclass import AbstractTrajectoryArray
+from .abstract_trajectory_array_dataclass import AbstractTrajectoryArray
 from .abstract_trajectory_features_bag_dataclass import AbstractTrajectoryFeaturesBag
 from .abstract_trajectory_stamped_features_bag_dataclass import (
     AbstractTrajectoryStampedFeaturesBag,
 )
 from .base_trajectory_dataclass import (
     BaseTrajectoryFeature,
-    BaseTrajectoryArray,
-    NestedBaseTrajectory,
+    BaseTrajectoryFeatureArray,
 )
