+++ tests/tests_extractor/tests_rosbag/test_rosbag_to_trajectory_features_bag.py
@@ -9,9 +9,8 @@ from trajectory_container_tools.extractor.rosbag_to_tct import (
     from_rosbag,
     extract_rosbag_feature,
 )
-from trajectory_container_tools.dataclasses.ros_msgs.primitive_dataclass import StdMsgsHeader
 
-from trajectory_container_tools.dataclasses.ros_msgs.non_trajectory_dataclass import (
+from trajectory_container_tools.dataclasses.ros_msgs.tf_dataclass import (
     Tf2MsgsTFMessage,
 )
 from trajectory_container_tools.dataclasses import (
@@ -21,7 +20,7 @@ from trajectory_container_tools.dataclasses import (
     RosStampedFeature,
     SensorMsgsLaserScan,
     SensorMsgsImu,
-    VescMsgsVescImuStamped,
+    StdMsgsHeader, VescMsgsVescImuStamped,
 )
 from trajectory_container_tools.temporal.timestamps import (
     TimestampCausalOrderingError,
