+++ tests/tests_utils/test_trajectory_dataclass_factory.py
@@ -5,8 +5,7 @@ import pytest
 from trajectory_container_tools.dataclasses.panda_dataframe_feature_dataclass import (
     BaseDataframeFeatureDataclass,
 )
-from trajectory_container_tools.dataclasses.ros_msgs.primitive_dataclass import StdMsgsHeader
-from trajectory_container_tools.dataclasses import RosStampedFeature
+from trajectory_container_tools.dataclasses import RosStampedFeature, StdMsgsHeader
 from trajectory_container_tools.utils.factory import (
     TrjDataClassFeatureSpecification,
     create_dataclass,
